+++ src/main/java/hs/kr/backend/devpals/global/jwt/JwtTokenValidator.java
@@ -53,9 +53,9 @@ public class JwtTokenValidator {
     /**
      * JWT 토큰에서 userId(PK) 추출
      */
-    public Integer getUserId(String jwt) {
+    public Long getUserId(String jwt) {
         String token = jwt.replace("Bearer ", "");
-        return Integer.parseInt(extractClaims(token).getSubject()); //  subject에서 userId(PK) 가져옴
+        return Long.parseLong(extractClaims(token).getSubject()); //  subject에서 userId(PK) 가져옴
     }
 
     /**
